refactor(recent-food): migrate to clean architecture with separated layers
- Create domain layer with pure types and repository interface
- Create infrastructure layer with Supabase implementation and Zod schemas
- Refactor application layer to use only interfaces and domain logic
- Remove direct Supabase and Zod usage from application layer
- Maintain backward compatibility and all existing APIs
- All tests passing (286 passed, 3 skipped)
