+++ src/modules/recent-food/application/recentFood.ts
@@ -1,102 +1,23 @@
-// Application layer for recent food operations, migrated from legacy controller
-// All error handling is done here, domain remains pure
-import { z } from 'zod'
-
-import { foodSchema } from '~/modules/diet/food/domain/food'
-import { recipeSchema } from '~/modules/diet/recipe/domain/recipe'
+// Application layer for recent food operations - pure business logic
 import type { Template } from '~/modules/diet/template/domain/template'
+import {
+  createRecentFoodInput,
+  type RecentFoodCreationParams,
+  type RecentFoodInput,
+  type RecentFoodRecord,
+  type RecentFoodRepository,
+  type RecentFoodType,
+} from '~/modules/recent-food/domain/recentFood'
+import {
+  supabaseRecentFoodRepository,
+  transformRowToTemplate,
+} from '~/modules/recent-food/infrastructure/supabaseRecentFoodRepository'
 import { showPromise } from '~/modules/toast/application/toastManager'
 import env from '~/shared/config/env'
 import { handleApiError } from '~/shared/error/errorHandler'
-import { parseWithStack } from '~/shared/utils/parseWithStack'
-import { removeDiacritics } from '~/shared/utils/removeDiacritics'
-import supabase from '~/shared/utils/supabase'
 
-const TABLE = 'recent_foods'
-
-// Database record type (replaces RecentFood domain entity)
-type RecentFoodRecord = {
-  id: number
-  user_id: number
-  type: 'food' | 'recipe'
-  reference_id: number
-  last_used: Date
-  times_used: number
-}
-
-// Input type for creating/updating recent foods
-type RecentFoodInput = {
-  user_id: number
-  type: 'food' | 'recipe'
-  reference_id: number
-  last_used?: Date
-  times_used?: number
-}
-
-// Creation params type (for backward compatibility)
-type RecentFoodCreationParams = Partial<RecentFoodRecord> &
-  Pick<RecentFoodRecord, 'user_id' | 'type' | 'reference_id'>
-
-// Database record schema for validation
-const recentFoodRecordSchema = z.object({
-  id: z.number(),
-  user_id: z.number(),
-  type: z.enum(['food', 'recipe']),
-  reference_id: z.number(),
-  last_used: z.coerce.date(),
-  times_used: z.number(),
-})
-
-/**
- * Creates a recent food input object (replaces createNewRecentFood)
- */
-export function createRecentFoodInput(
-  params: RecentFoodCreationParams,
-): RecentFoodInput {
-  return {
-    user_id: params.user_id,
-    type: params.type,
-    reference_id: params.reference_id,
-    last_used: new Date(),
-    times_used: (params.times_used ?? 0) + 1,
-  }
-}
-
-// Schema for the enhanced database function response
-const enhancedRecentFoodRowSchema = z
-  .object({
-    recent_food_id: z.number(),
-    user_id: z.number(),
-    type: z.enum(['food', 'recipe']),
-    reference_id: z.number(),
-    last_used: z.coerce.date(),
-    times_used: z.number(),
-    template_id: z.number(),
-    template_name: z.string(),
-    template_ean: z.string().nullable(),
-    template_source: z.unknown(),
-    template_macros: z.unknown(),
-    template_owner: z.number().nullable(),
-    template_items: z.unknown(),
-    template_prepared_multiplier: z.number().nullable(),
-  })
-  .strip()
-
-// Helper function to safely get recipe fields
-function getRecipeFields(row: z.infer<typeof enhancedRecentFoodRowSchema>) {
-  if (row.type !== 'recipe') {
-    throw new Error('Expected recipe type but got food')
-  }
-
-  const owner = row.template_owner
-  const preparedMultiplier = row.template_prepared_multiplier
-
-  if (owner === null || preparedMultiplier === null) {
-    throw new Error('Recipe fields cannot be null')
-  }
-
-  return { owner, preparedMultiplier }
-}
+// Default repository implementation (can be swapped for testing)
+const recentFoodRepository: RecentFoodRepository = supabaseRecentFoodRepository
 
 /**
  * Fetches a recent food by user, type and reference ID.
@@ -107,24 +28,14 @@ function getRecipeFields(row: z.infer<typeof enhancedRecentFoodRowSchema>) {
  */
 export async function fetchRecentFoodByUserTypeAndReferenceId(
   userId: number,
-  type: 'food' | 'recipe',
+  type: RecentFoodType,
   referenceId: number,
 ): Promise<RecentFoodRecord | null> {
-  try {
-    const { data, error } = await supabase
-      .from(TABLE)
-      .select('*')
-      .eq('user_id', userId)
-      .eq('type', type)
-      .eq('reference_id', referenceId)
-    if (error !== null) throw error
-
-    const records = parseWithStack(recentFoodRecordSchema.array(), data)
-    return records.at(0) ?? null
-  } catch (error) {
-    handleApiError(error)
-    return null
-  }
+  return recentFoodRepository.fetchByUserTypeAndReferenceId(
+    userId,
+    type,
+    referenceId,
+  )
 }
 
 /**
@@ -141,46 +52,14 @@ export async function fetchUserRecentFoods(
   search?: string,
 ): Promise<readonly Template[]> {
   try {
-    const normalizedSearch =
-      search?.trim() !== undefined && search.trim() !== ''
-        ? removeDiacritics(search.trim())
-        : undefined
-    const response = await supabase.rpc('search_recent_foods_with_names', {
-      p_user_id: userId,
-      p_search_term: normalizedSearch ?? null,
-      p_limit: limit,
-    })
-    if (response.error !== null) throw response.error
-
-    // Transform the enhanced result to Template objects
-    const validatedRows = parseWithStack(
-      enhancedRecentFoodRowSchema.array(),
-      response.data,
+    const rawRows = await recentFoodRepository.fetchUserRecentFoodsRaw(
+      userId,
+      limit,
+      search,
     )
-    const templates = validatedRows.map((row) => {
-      // Create the appropriate Template object based on type
-      if (row.type === 'food') {
-        return parseWithStack(foodSchema, {
-          id: row.template_id,
-          name: row.template_name,
-          ean: row.template_ean,
-          source: row.template_source,
-          macros: row.template_macros,
-          __type: 'Food',
-        })
-      } else {
-        const { owner, preparedMultiplier } = getRecipeFields(row)
-        return parseWithStack(recipeSchema, {
-          id: row.template_id,
-          name: row.template_name,
-          owner,
-          items: row.template_items,
-          prepared_multiplier: preparedMultiplier,
-          __type: 'Recipe',
-        })
-      }
-    })
 
+    // Transform raw data to Template objects
+    const templates = rawRows.map((row) => transformRowToTemplate(row))
     return templates
   } catch (error) {
     handleApiError(error)
@@ -198,21 +77,7 @@ export async function insertRecentFood(
 ): Promise<RecentFoodRecord | null> {
   try {
     return await showPromise(
-      (async () => {
-        const insertData = {
-          user_id: recentFoodInput.user_id,
-          type: recentFoodInput.type,
-          reference_id: recentFoodInput.reference_id,
-          last_used: recentFoodInput.last_used ?? new Date(),
-          times_used: recentFoodInput.times_used ?? 1,
-        }
-        const { data, error } = await supabase
-          .from(TABLE)
-          .insert(insertData)
-          .select()
-        if (error !== null) throw error
-        return parseWithStack(recentFoodRecordSchema, data[0])
-      })(),
+      recentFoodRepository.insert(recentFoodInput),
       {
         loading: 'Salvando alimento recente...',
         success: 'Alimento recente salvo com sucesso',
@@ -238,22 +103,7 @@ export async function updateRecentFood(
 ): Promise<RecentFoodRecord | null> {
   try {
     return await showPromise(
-      (async () => {
-        const updateData = {
-          user_id: recentFoodInput.user_id,
-          type: recentFoodInput.type,
-          reference_id: recentFoodInput.reference_id,
-          last_used: recentFoodInput.last_used ?? new Date(),
-          times_used: recentFoodInput.times_used ?? 1,
-        }
-        const { data, error } = await supabase
-          .from(TABLE)
-          .update(updateData)
-          .eq('id', recentFoodId)
-          .select()
-        if (error !== null) throw error
-        return parseWithStack(recentFoodRecordSchema, data[0])
-      })(),
+      recentFoodRepository.update(recentFoodId, recentFoodInput),
       {
         loading: 'Atualizando alimento recente...',
         success: 'Alimento recente atualizado com sucesso',
@@ -276,20 +126,12 @@ export async function updateRecentFood(
  */
 export async function deleteRecentFoodByReference(
   userId: number,
-  type: 'food' | 'recipe',
+  type: RecentFoodType,
   referenceId: number,
 ): Promise<boolean> {
   try {
-    await showPromise(
-      (async () => {
-        const { error } = await supabase
-          .from(TABLE)
-          .delete()
-          .eq('user_id', userId)
-          .eq('type', type)
-          .eq('reference_id', referenceId)
-        if (error !== null) throw error
-      })(),
+    return await showPromise(
+      recentFoodRepository.deleteByReference(userId, type, referenceId),
       {
         loading: 'Removendo alimento recente...',
         success: 'Alimento recente removido com sucesso',
@@ -297,11 +139,17 @@ export async function deleteRecentFoodByReference(
       },
       { context: 'user-action', audience: 'user' },
     )
-    return true
   } catch (error) {
     handleApiError(error)
     return false
   }
 }
 
-export {}
+// Re-export domain functions for convenience
+export { createRecentFoodInput }
+export type {
+  RecentFoodCreationParams,
+  RecentFoodInput,
+  RecentFoodRecord,
+  RecentFoodType,
+}
